Just ran depth-13. 60% of 1LLL has been tackled.

+++ solverc/shiftcube/src/alg.h
@@ -6,8 +6,8 @@
 
 
 typedef struct {
-    size_t size;
-    size_t length;
+    uint8_t size;
+    uint8_t length;
     move_t *moves;
 } alg_s;
